EL PR change requests
