+++ realm/realm-library/src/testUtils/kotlin/io/realm/entities/MixedNotIndexed.kt
@@ -1,0 +1,12 @@
+package io.realm.entities
+
+import io.realm.Mixed
+import io.realm.RealmObject
+
+open class MixedNotIndexed : RealmObject() {
+    companion object {
+        const val FIELD_MIXED = "mixed"
+    }
+
+    var mixed: Mixed? = null
+}
